+++ de.walware.statet.r.ui/src/de/walware/statet/r/internal/ui/datafilterview/TextClient.java
@@ -94,7 +94,7 @@ public class TextClient extends FilterClient {
 		
 		@Override
 		public Object execute(final ExecutionEvent event) throws ExecutionException {
-			final List<String> values = new ArrayList<String>(fAvailableValues.getLength() - fSelectedValueSet.size());
+			final List<String> values = new ArrayList<String>((int) fAvailableValues.getLength() - fSelectedValueSet.size());
 			for (int i = 0; i < fAvailableValues.getLength(); i++) {
 				final String value = fAvailableValues.get(i);
 				if (!fSelectedValueSet.contains(value)) {
@@ -300,7 +300,7 @@ public class TextClient extends FilterClient {
 	
 	@Override
 	protected boolean updateLayout() {
-		return updateLayout(fValueListViewer, fAvailableValues.getLength());
+		return updateLayout(fValueListViewer, (int) fAvailableValues.getLength());
 	}
 	
 	@Override
